Produce two libraries, libnvme and libnvme-spec
Make the build process create two libraries, libnvme and libnvme-spec.
libnvme-spec is a subset of libnvme, which means that, existing
applications does not need to change - they can continue to link with
libnvme and get all the goodies. libnvme-spec is for applications and
libraries that are not interested in linking with the parts of libnvme
that are Linux kernel specific.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>
[rebase to current master; Jeremy Kerr <jk@codeconstruct.com.au>]